Add full testing of controllers and services

+++ src/main/kotlin/net/ddellspe/escapegenai/repository/InvoiceRepository.kt
@@ -4,4 +4,4 @@ import net.ddellspe.escapegenai.model.Invoice
 import org.springframework.data.repository.ListCrudRepository
 import org.springframework.stereotype.Repository
 
-@Repository interface InvoiceRepository : ListCrudRepository<Invoice, Int> {}
+@Repository interface InvoiceRepository : ListCrudRepository<Invoice, Long> {}
